packages/content-pipeline/pipeline/config.py
"""환경변수 및 설정"""

import os
from dotenv import load_dotenv

load_dotenv()

ANTHROPIC_API_KEY = os.environ.get("ANTHROPIC_API_KEY", "")
SUPABASE_URL = os.environ.get("SUPABASE_URL", "")
SUPABASE_SERVICE_KEY = os.environ.get("SUPABASE_SERVICE_KEY", "")

MODEL_NAME = "claude-sonnet-4-6"
MAX_RETRIES = 3
