packages/content-pipeline/pyproject.toml
[project]
name = "suji-content-pipeline"
version = "0.1.0"
description = "수지 수학 학습앱 콘텐츠 파이프라인 — PDF에서 개념 추출 → AI 문제 생성 → DB 저장"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pypdf>=4.0",
    "sympy>=1.13",
    "supabase>=2.0",
    "typer[all]>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
suji-pipeline = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]
